rename : 폴더 병합 및 필요없는 파일 제거
